cycle counts reset every frame

+++ include/ppu.h
@@ -1,6 +1,6 @@
 #include <SDL2/SDL.h>
 
-void ppu_tick_to(unsigned long long cycle);
+void ppu_tick_to(int cycle);
 void ppu_setup(unsigned char *chrom, unsigned char *vram, unsigned char *palette, unsigned char *oam2, int *nmi, SDL_Texture *texture, SDL_Renderer *renderer, unsigned char *pixels);
 unsigned char ppu_read(unsigned short addr);
 void ppu_write(unsigned short addr, unsigned char data);
